Add tweaks to manipulate voodoo doll behavior
Some PWAD maps use voodoo dolls as "macros" to, e.g. manipulate multiple
parts of the map from one switch press, exit the level when some monster
is killed, etc.

The vanilla setup for this requires the voodoo doll to take damage from some
source, usually a barrel or a DEHACKED object that acts as an infinite barrel,
which pushes the voodoo doll across lines to activate. However, this also
means that the player will take some damage inexplicably; this is usually
handled by additional healing items in the path of the voodoo doll, but that
still won't save a player at 1% health that triggers it.

So instead, there are now explicit mapthing tweaks to make voodoo dolls not
transfer their damage or item collection to the main player. This allows these
voodoo "macros" to function without having a visible effect on the player.

+++ src/doom/p_inter.c
@@ -400,6 +400,8 @@ P_TouchSpecialThing
     // [AP] transfer voodoo doll special touch to real player
     if (toucher->type == MT_FAKEPLAYER)
     {
+      if (toucher->spawnpoint.options & APMTF_VOODOO_NOITEMS)
+        return; // This voodoo doll is used for things like level movement/scripts!
       if (special->sprite == SPR_LVST)
         return; // Voodoo dolls can't exit to hub
       if (!player->mo || player->mo->health <= 0)
@@ -1112,6 +1114,8 @@ P_DamageMobj
     // [AP] transfer voodoo doll damage to real player
     if (player && target->type == MT_FAKEPLAYER)
     {
+      if (target->spawnpoint.options & APMTF_VOODOO_NODAMAGE)
+        return; // This voodoo doll is used for things like level movement/scripts!
       if (!player->mo || player->mo->health <= 0)
         return; // Actual player is dead, don't kill them further
       target = player->mo;
